@@ -12,7 +12,6 @@
     "dev": "next dev",
     "prebuild": "echo 'Building Sanity to public/admin' && cd admin && yarn && npx @sanity/cli build ../public/admin -y && echo 'Done'",
     "build": "next build",
-    "postbuild": "next-sitemap --config next-sitemap.config.js",
     "start": "next start",
     "sanity": "cd admin && sanity start",
     "lint": "next lint"
@@ -20,38 +19,36 @@
   "dependencies": {
     "@cloudinary/react": "^1.11.2",
     "@cloudinary/url-gen": "^1.11.2",
-    "@next/bundle-analyzer": "^13.4.19",
+    "@next/bundle-analyzer": "^13.5.3",
     "@portabletext/react": "^3.0.7",
     "@react-spring/three": "^9.7.3",
-    "@react-three/drei": "^9.83.9",
-    "@react-three/fiber": "^8.14.1",
+    "@react-three/drei": "^9.86.0",
+    "@react-three/fiber": "^8.14.3",
     "@sanity/client": "^6.4.12",
     "@sanity/image-url": "^1.0.2",
     "@vercel/analytics": "^1.0.2",
     "age-calculator": "^1.0.0",
     "color": "^4.2.3",
-    "groq": "^3.16.4",
+    "groq": "^3.16.7",
     "hamburger-react": "^2.5.0",
-    "next": "^13.4.19",
+    "next": "^13.5.3",
     "next-pwa": "^5.6.0",
-    "next-sanity": "^5.5.3",
-    "next-sitemap": "^4.2.3",
+    "next-sanity": "^5.5.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-html-comment": "^2.0.16",
     "react-icons": "^4.11.0",
     "react-palette": "^1.0.2",
-    "react-spring": "^9.7.2",
+    "react-spring": "^9.7.3",
     "react-use-measure": "^2.1.1",
-    "sass": "^1.67.0",
+    "sass": "^1.68.0",
     "seedrandom": "^3.0.5",
-    "three": "^0.156.1",
-    "three-stdlib": "^2.26.0"
+    "three": "^0.157.0",
+    "three-stdlib": "^2.26.9"
   },
   "devDependencies": {
-    "@babel/core": "^7.22.17",
-    "eslint": "^8.49.0",
-    "eslint-config-next": "^13.4.19",
+    "@babel/core": "^7.23.0",
+    "eslint": "^8.50.0",
+    "eslint-config-next": "^13.5.3",
     "eslint-config-prettier": "^9.0.0",
     "prettier": "^3.0.3",
     "r3f-perf": "^7.1.2",
