@@ -1,5 +1,6 @@
-import { Component } from 'react';
-import HTMLComment from 'react-html-comment';
+'use client';
+
+import { useEffect, useRef } from 'react';
 
 const str =
   '-\n\
@@ -21,8 +22,17 @@ const str =
                                                .:..: -:               \n\
 -';
 
-export default class Signature extends Component {
-  render() {
-    return <HTMLComment text={str} />;
-  }
+export default function Signature() {
+  const ref = useRef(null);
+  const commentAdded = useRef(false);
+
+  useEffect(() => {
+    if (ref.current && !commentAdded.current) {
+      const comment = document.createComment(str);
+      ref.current.parentNode.insertBefore(comment, ref.current);
+      commentAdded.current = true;
+    }
+  }, []);
+
+  return <div ref={ref}></div>;
 }
