@@ -1,5 +1,3 @@
-'use client';
-
 /*
 Auto-generated by: https://github.com/pmndrs/gltfjsx
 Command: npx gltfjsx@6.2.10 C:\Users\Ruben\Downloads\RP-Logo.glb -T --shadows
@@ -8,7 +6,7 @@ Command: npx gltfjsx@6.2.10 C:\Users\Ruben\Downloads\RP-Logo.glb -T --shadows
 import { useState, useEffect } from 'react';
 import { useGLTF } from '@react-three/drei';
 import { a, useSpring } from '@react-spring/three';
-import { useColor } from '../lib/themes';
+import { useColor } from '../app/_themes/client';
 
 export function RpLogo(props) {
   const [scale, setScale] = useState(Math.min(1, window.screen.width / 600));
@@ -28,7 +26,7 @@ export function RpLogo(props) {
     config: { duration: 20000 },
     loop: true,
     from: { y: -0.065 },
-    to: { y: (Math.PI * 2) - 0.065 }
+    to: { y: Math.PI * 2 - 0.065 }
   });
 
   useEffect(() => {
