@@ -44,9 +44,9 @@ const getThemeUrl = async () => {
     },
     prioritize
   }`;
-  const themes = await client.fetch(query);
+  const themes = await server.fetch(query);
   for (const theme of themes) {
-    if(!theme.prioritize) {
+    if (!theme.prioritize) {
       data.urls.push(...theme.urls);
     } else {
       data.prioritized.push(...theme.urls);
@@ -63,7 +63,7 @@ const getThemeUrl = async () => {
     console.log(data[index].message);
   }
   return data[index].image;
-}
+};
 
 const GetAboutPageData = async () => {
   const query = `
@@ -74,6 +74,6 @@ const GetAboutPageData = async () => {
   }`;
   const content = await client.fetch(query);
   return content;
-}
+};
 
 export { client, server, getThemeUrl, GetAboutPageData };
