@@ -9,54 +9,65 @@ const withPWA = require('next-pwa')({
 });
 
 const nextConfig = {
-    swcMinify: true,
-    transpilePackages: ['@acme/ui', 'lodash-es'],
-    modularizeImports: {
-      'react-bootstrap': {
-        transform: 'react-bootstrap/{{member}}'
-      },
-      lodash: {
-        transform: 'lodash/{{member}}',
-        preventFullImport: true
-      }
+  swcMinify: true,
+  transpilePackages: ['@acme/ui', 'lodash-es'],
+  modularizeImports: {
+    'react-bootstrap': {
+      transform: 'react-bootstrap/{{member}}'
     },
-    async rewrites() {
-      return [
-        {
-          source: '/admin/:path*',
-          destination:
-            process.env.NODE_ENV === 'development'
-              ? 'http://localhost:3333/admin/:path*'
-              : '/admin/index.html'
-        }
-      ];
-    },
-    i18n: {
-      locales: ['en'],
-      defaultLocale: 'en'
-    },
-    reactStrictMode: true,
-    images: {
-      domains: ['res.cloudinary.com']
+    lodash: {
+      transform: 'lodash/{{member}}',
+      preventFullImport: true
     }
-}
+  },
+  async rewrites() {
+    return [
+      {
+        source: '/admin/:path*',
+        destination:
+          process.env.NODE_ENV === 'development'
+            ? 'http://localhost:3333/admin/:path*'
+            : '/admin/index.html'
+      }
+    ];
+  },
+  i18n: {
+    locales: ['en'],
+    defaultLocale: 'en'
+  },
+  reactStrictMode: true,
+  images: {
+    domains: ['res.cloudinary.com']
+  }
+};
 
-const KEYS_TO_OMIT = ['webpackDevMiddleware', 'configOrigin', 'target', 'analyticsId', 'webpack5', 'amp', 'assetPrefix']
+const KEYS_TO_OMIT = [
+  'webpackDevMiddleware',
+  'configOrigin',
+  'target',
+  'analyticsId',
+  'webpack5',
+  'amp',
+  'assetPrefix'
+];
 
 module.exports = (_phase, { defaultConfig }) => {
-  const plugins = [[withPWA], [withBundleAnalyzer, {}]]
+  const plugins = [[withPWA], [withBundleAnalyzer, {}]];
 
-  const wConfig = plugins.reduce((acc, [plugin, config]) => plugin({ ...acc, ...config }), {
-    ...defaultConfig,
-    ...nextConfig,
-  })
+  const wConfig = plugins.reduce(
+    (acc, [plugin, config]) => plugin({ ...acc, ...config }),
+    {
+      ...defaultConfig,
+      ...nextConfig
+    }
+  );
 
-  const finalConfig = {}
+  const finalConfig = {};
   Object.keys(wConfig).forEach((key) => {
     if (!KEYS_TO_OMIT.includes(key)) {
-      finalConfig[key] = wConfig[key]
+      finalConfig[key] = wConfig[key];
     }
-  })
+  });
 
-  return finalConfig
-}
+  return finalConfig;
+};
